chatbot_app/requirements.txt
groq==0.4.1
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
pytest==8.1.1
rich==13.7.1
html2text==2024.2.26